+++ internal/authentication/oauth_test.go
@@ -1,4 +1,4 @@
-package common
+package authentication
 
 import (
 	"testing"
