+++ internal/core/web.go
@@ -1,0 +1,1 @@
+package core
