--- internal/portal/web.go
@@ -1,1 +1,0 @@
-package portal
